popular-packages/requirements.txt
python-dotenv==1.1.1
sendgrid==6.12.5
twilio==9.8.3
requests==2.32.5
